Moved user defined application and api logic into the oauth package
